+++ src/test/scala/sigmastate/utxo/examples/DemurrageExampleSpecification.scala
@@ -62,14 +62,14 @@ class DemurrageExampleSpecification extends SigmaTestingCommons {
         | val out = OUTPUTS(outIdx)
         |
         | val c1 = allOf(Coll(
-        |   HEIGHT >= SELF.R3[(Long, Coll[Byte])].get._1 + demurragePeriod,
-        |   SELF.value - demurrageCoeff * SELF.bytes.size * (HEIGHT - SELF.R3[(Long, Coll[Byte])].get._1) <= 0
+        |   HEIGHT >= SELF.R3[(Int, Coll[Byte])].get._1 + demurragePeriod,
+        |   SELF.value - demurrageCoeff * SELF.bytes.size * (HEIGHT - SELF.R3[(Int, Coll[Byte])].get._1) <= 0
         | ))
         |
         | val c2 = allOf(Coll(
-        |   HEIGHT >= SELF.R3[(Long, Coll[Byte])].get._1 + demurragePeriod,
-        |   out.R3[(Long, Coll[Byte])].get._1 == HEIGHT,
-        |   out.value >= SELF.value - demurrageCoeff * SELF.bytes.size * (HEIGHT - SELF.R3[(Long, Coll[Byte])].get._1),
+        |   HEIGHT >= SELF.R3[(Int, Coll[Byte])].get._1 + demurragePeriod,
+        |   out.R3[(Int, Coll[Byte])].get._1 == HEIGHT,
+        |   out.value >= SELF.value - demurrageCoeff * SELF.bytes.size * (HEIGHT - SELF.R3[(Int, Coll[Byte])].get._1),
         |   out.propositionBytes == SELF.propositionBytes
         | ))
         |
